refactored main() out from bulk_extractor.cpp, so that we can do end-to-end testing on the executable
